+++ bin/worker.js
@@ -1,24 +1,25 @@
 "use strict"
 
-const {parentPort} = require("worker_threads")
 const o = require("../ospec")
+const {
+	parent, sendMessage, decodeMessage,
+	loadWithImport, loadWithRequire, useModule
+} = require("./compat")
 
 function cloneErr({name, stack, message}) {
 	return {name, stack, message}
 }
 
-parentPort.on("message", async function(path) {
+const load = useModule ? loadWithImport : loadWithRequire
+
+parent.on("message", async function(path) {
 	try {
-		await import(path)
+		await load(decodeMessage(path))
 		o.run(function(results) {
-			// JSON.stringify()/.parse() is faster than the structured copy
-			// algorithm, especially for large objects. Also, postMessage
-			// seems not to like the results as they are handed by ospec.
-			// It hangs forever unless the objects are first copied
-			// manually.
-			// We also compress the results, since most tests are expected to
+			// We compress the results, since most tests are expected to
 			// pass on a typical run.
-			parentPort.postMessage(JSON.stringify(
+			sendMessage(
+				parent,
 				results.reduce(
 					(acc, r) => {
 						if (r.pass) {
@@ -34,17 +35,20 @@ parentPort.on("message", async function(path) {
 					},
 					{pass:0, fail:[]}
 				)
-			))
+			)
 		})
 	} catch(e) {
-		parentPort.postMessage(JSON.stringify({
-			pass: 0,
-			fail:[{
-				pass: false,
-				context: e.message,
-				message: e.stack,
-				error: cloneErr(e)
-			}]
-		}))
+		sendMessage(
+			parent,
+			{
+				pass: 0,
+				fail: [{
+					pass: false,
+					context: e.message,
+					message: e.stack,
+					error: cloneErr(e)
+				}]
+			}
+		)
 	}
 })